+++ tests/Unit/Casting/ExtractTest.php
@@ -6,6 +6,8 @@ namespace Nandan108\DtoToolkit\Tests\Unit\Casting;
 
 use Nandan108\DtoToolkit\CastTo;
 use Nandan108\DtoToolkit\Core\FullDto;
+use Nandan108\DtoToolkit\Core\ProcessingContext;
+use Nandan108\DtoToolkit\Exception\Process\ExtractionException;
 use Nandan108\DtoToolkit\Tests\Traits\CanTestCasterClassesAndMethods;
 use Nandan108\PropAccess\PropAccess;
 use PHPUnit\Framework\TestCase;
@@ -34,17 +36,19 @@ final class ExtractTest extends TestCase
         ]);
         $dto->setGetterVal(['D', 'E', 'F']);
 
-        // Extract through objects, via props or getters
-        $this->casterTest(new CastTo\Extract('foo[propVal.1, context.baz]'), ['foo' => $dto], ['B', 'qux']);
-        $this->casterTest(new CastTo\Extract('dto.getterVal.1'), ['dto' => $dto], 'E');
-        $this->casterTest(new CastTo\Extract('propVal.2.2'), $dto, 'C');
-        $this->casterTest(new CastTo\Extract('propVal.2.2'), $dto, 'C');
+        ProcessingContext::wrapProcessing($dto, function () use ($dto) {
+            // Extract through objects, via props or getters
+            $this->casterTest(new CastTo\Extract('foo[propVal.1, context.baz]'), ['foo' => $dto], ['B', 'qux']);
+            $this->casterTest(new CastTo\Extract('dto.getterVal.1'), ['dto' => $dto], 'E');
+            $this->casterTest(new CastTo\Extract('propVal.2.2'), $dto, 'C');
+            $this->casterTest(new CastTo\Extract('propVal.2.2'), $dto, 'C');
 
-        // Extract from ArrayAccess with invalid path
-        $this->casterTest(new CastTo\Extract('propVal.2.!3'), $dto, \Nandan108\DtoToolkit\Exception\Process\ExtractionException::class);
+            // Extract from ArrayAccess with invalid path
+            $this->casterTest(new CastTo\Extract('propVal.2.!3'), $dto, ExtractionException::class);
 
-        // Extract with invalid path input
-        $this->casterTest(new CastTo\Extract('foo.bar.!3'), ['foo' => ['bar' => ['a', 'b', 'c']]], \Nandan108\DtoToolkit\Exception\Process\ExtractionException::class);
+            // Extract with invalid path input
+            $this->casterTest(new CastTo\Extract('foo.bar.!3'), ['foo' => ['bar' => ['a', 'b', 'c']]], ExtractionException::class);
+        });
     }
 
     public function testExtractFailsWithInvalidPath(): void